changed BitOperator to only return the necessery bytes ignoring the two's complement representation where an additional byte is added when the MSB is set.

+++ dsl/src/main/groovy/at/fabianachammer/pdusend/dsl/PduSendVocabulary.groovy
@@ -60,7 +60,7 @@ class PduSendVocabulary {
 
 			else if(allowedNumberTypes.any{n -> n == input.class}){
 				if(fieldClass == byte[].class){
-					pdu."$it" = BitOperator.split(input, pdu."$it".length)
+					pdu."$it" = BitOperator.split(input as BigInteger, pdu."$it".length)
 				}
 
 				else if(pdu."$it" instanceof DataUnit){
